DAOS-11266 test: Extend CommandWithParameters from DdbCommandBase
Also consolidate ddb_utils.py and ddb_utils_base.py

Skip-unit-tests: true
Skip-fault-injection-test: true
Test-tag: pr
Required-githooks: true
